refactor(app): extract plugin command dispatch (phase 6a)
Move update_plugin_state_snapshot, handle_plugin_command, and the
~25 plugin-specific handler methods (handle_save_buffer_to_path,
handle_load_plugin, handle_unload_plugin, handle_reload_plugin,
handle_list_plugins, handle_execute_action, handle_execute_actions,
handle_get_buffer_text, handle_set_editor_mode,
handle_get_line_start_position, handle_get_line_end_position,
handle_get_buffer_line_count, handle_scroll_to_line_center,
handle_scroll_buffer_to_line) out of mod.rs into a new
app/plugin_dispatch.rs sibling. ~2300 lines.

This is the biggest single extraction in the refactor: the plugin
dispatch + handler cluster was 2300 lines of plugin-API surface that
had no business living inside the central Editor module file.

mod.rs goes from 9505 -> 7218 lines (-2287). 2026 tests still pass.
